Move new answer form to its own controller
