+++ frontend/app/models/answer.js
@@ -1,15 +1,8 @@
-import Ember from 'ember';
 import DS from 'ember-data';
 
-export default DS.Model.extend(Ember.Validations.Mixin, {
+export default DS.Model.extend({
   question: DS.belongsTo('question', { async: true }),
   user: DS.belongsTo('user', { async: true }),
 
-  body: DS.attr('string'),
-
-  validations: {
-    body: {
-      length: { minimum: 30, maximum: 10000 }
-    }
-  }
+  body: DS.attr('string')
 });
