+++ frontend/app/routes/questions/show.js
@@ -13,22 +13,6 @@ export default Ember.Route.extend({
   actions: {
     didTransition: function() {
       this.get('controller').set('isEditing', false);
-    },
-
-    saveAnswer: function() {
-      var question = this.currentModel;
-      var answer = this.controller.get('answer');
-      var _this = this;
-
-      answer.set('question', question);
-
-      answer.save().then(function() {
-        question.get('answers').pushObject(answer);
-        _this.wuphf.success('Thanks for answering!', 3000);
-
-        var newAnswer = _this.store.createRecord('answer');
-        _this.controller.set('answer', newAnswer);
-      });
     }
   }
 });
