+++ frontend/app/templates/answers/new.hbs
@@ -1,0 +1,6 @@
+<div class="answer-form">
+  {{#form-for controller action='save'}}
+    {{input body as="text"}}
+    {{submit 'Post Answer' class='button'}}
+  {{/form-for}}
+</div>
