+++ frontend/app/templates/questions/show.hbs
@@ -59,12 +59,7 @@
 <h3>Your Answer</h3>
 
 {{#if session.isAuthenticated}}
-  <div class="answer-form">
-    {{#form-for answer action='saveAnswer'}}
-      {{input body as="text"}}
-      {{submit 'Post Answer' class='button'}}
-    {{/form-for}}
-  </div>
+  {{render 'answers.new' answer}}
 {{else}}
   {{link-to 'Log in to post an answer' 'login'}}
 {{/if}}
