+++ frontend/tests/unit/controllers/answers/new-test.js
@@ -1,0 +1,12 @@
+import { test, moduleFor } from 'ember-qunit';
+
+moduleFor('controller:answers/new', 'AnswersNewController', {
+  // Specify the other units that are required for this test.
+  // needs: ['controller:foo']
+});
+
+// Replace this with your real tests.
+test('it exists', function() {
+  var controller = this.subject();
+  ok(controller);
+});
